requirements.txt
google-adk
litellm
python-dotenv
pydantic
chromadb
langchain-openai
sentence-transformers
PyMuPDF
google-generativeai
